+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "des"
+name = "esb"
 version = "0.1.0"
 description = ""
 authors = ["Rowan Molony <rowan.molony@codema.ie>"]
@@ -15,6 +15,7 @@ matplotlib = "^3.3.3"
 pandas = ">=1.1.5"
 notebook = "^6.2.0"
 geopandas = "^0.8.2"
+pygeos = "^0.9"
 
 [tool.poetry.dev-dependencies]
 black = "^20.8b1"
